Modeles in tests/ProcessMaker/Models moved to tests/ProcessMaker/Test/Model
