+++ tests/Feature/Engine/EngineTestCase.php
@@ -3,10 +3,10 @@ namespace Tests\Feature\Engine;
 
 use PHPUnit\Framework\TestCase;
 use ProcessMaker\Bpmn\TestEngine;
-use ProcessMaker\Models\CallActivity;
-use ProcessMaker\Models\ExecutionInstanceRepository;
-use ProcessMaker\Models\ProcessRepository;
-use ProcessMaker\Models\TokenRepository;
+use ProcessMaker\Test\Models\CallActivity;
+use ProcessMaker\Test\Models\ExecutionInstanceRepository;
+use ProcessMaker\Test\Models\ProcessRepository;
+use ProcessMaker\Test\Models\TokenRepository;
 use ProcessMaker\Nayra\Bpmn\Lane;
 use ProcessMaker\Nayra\Bpmn\LaneSet;
 use ProcessMaker\Nayra\Bpmn\Models\Activity;
@@ -87,7 +87,7 @@ use ProcessMaker\Nayra\Contracts\Repositories\ProcessRepositoryInterface;
 use ProcessMaker\Nayra\Contracts\Repositories\TokenRepositoryInterface;
 use ProcessMaker\Nayra\Engine\ExecutionInstance;
 use ProcessMaker\Nayra\Factory;
-use ProcessMaker\Test\FormalExpression;
+use ProcessMaker\Test\Models\FormalExpression;
 
 /**
  * Test transitions
